Allow patches to be specified in .core files
Previously all patches in the 'patches' directory of a core root
were applied automatically. This gets problematic when we want to
have several cores in the same directory but apply a different set
of patches for each of them.

This patch makes it possible to specify the patches to apply as
a list in the 'patches' option in the 'main' section. Note that
the old behaviour still exists as well.