vivado: Make our run the default run
When opening the fusesoc-generated project in the Vivado GUI it shows a
new and empty synth_1 and impl_1 run as default instead of the one
already finished.

This set uses the runs created by default in Vivado's create_project,
avoiding the manual creation of the runs. (I've left the code to do so
in the TCL file as a starting point for manual adjustment should
multiple runs be required.)