Use YAML CSafeLoader/CSafeDumper if available
The pyyaml package is able to use libyaml for parsing and dumping YAML
files. In my tests, this gives a speedup of at least 10x when parsing
YAML files. The downside: pyyaml uses this library as an optional
dependency because it requires a C compilation setup and a native
library to be present. To ensure that the code always works, we
optionally use the C implementation, and fall back to the pure Python
implementation.

To make this change slightly cleaner, this patch centralizes all yaml
loading/dumping into three functions in fusesoc.utils.