Add support for specifying library version
This commit adds a new option sync-version to the config file and
the library subcommand for the cli tool. This option is passed through
to the provider used for a library to specify a version of the library
to use.
This change also updates the git provider to use the sync-version to
checkout a commitish on a git-provided library. This libraries to be
fixed at certian points, or follow specific branches (such as nightly or
stable)