CI: Run test in forked repo context
Right now, we run pull request CI jobs in the context of the
olofk/fusesoc repository. We did this to be able to create Check runs,
which show the pytest results in the GitHub UI. There are multiple
problems with that:

First, the runs actually didn't contain any data. That's probably
fixable if we found the right configuration regarding test result files
and artifacts.

Second, and more importantly, it seems that the "old" test code was run,
i.e. pytest didn't pick up test changes done in the PR. That's hard to
believe, but also not very useful, as PRs often modify code and tests at
the same time. There might be some misconfiguration going on somewhere,
but I didn't debug it fully.

Finally, the test result reporting action we used is actually rather
slow (significantly slower than the test runs themselves), as it creates
and then runs a Docker container.

In this commit, we try a different approach and annotate the source code
with test results.