Allow setting tags on files
This does not affect FuseSoC, but allows Edalize backends to take
different decisions depending on which tags that are set. Use-cases
for this is to mark HDL source files as simulation files to avoid
having them being sent to synthesis, and marking which tool in a flow
that should consume a particular TCL file.